Remove a bunch of VC7.1 compiler warnings
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@116 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfImage.cpp
@@ -147,7 +147,7 @@ void PdfImage::LoadFromFile( const char* pszFilename )
             break;
     }
     
-    this->SetImageData( m_rRect.Width(), m_rRect.Height(), 8 , szBuffer, lLen ); // 8 bits per component
+    this->SetImageData( (unsigned int)m_rRect.Width(), (unsigned int)m_rRect.Height(), 8 , szBuffer, lLen ); // 8 bits per component
 
     (void) jpeg_destroy_decompress(&cinfo);
